requirements.txt
Flask==3.1.0
MarkupSafe==2.1.5
gunicorn==21.2.0
python-dotenv==1.1.0
flask-talisman==1.0.0
requests==2.32.4
Flask-WTF==1.2.2